requirements.txt
flask>=2.0.1
requests>=2.26.0
gitpython>=3.1.18
pandas>=1.3.3
python-dateutil>=2.8.2
